requirements.txt
anyascii==0.3.0
asgiref==3.4.1
beautifulsoup4==4.9.3
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
cryptography==35.0.0
Django==3.2.8
django-crispy-forms==1.13.0
django-el-pagination==3.3.0
django-filebrowser-no-grappelli2==3.8.4
django-filter==2.4.0
django-js-asset==1.2.2
django-modelcluster==5.2
django-phonenumber-field==5.2.0
django-tinymce==3.3.0
django-widget-tweaks==1.4.9
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
ffmpeg==1.4
html5lib==1.1
idna==3.3
l18n==2020.6.1
Markdown==3.3.4
mysqlclient==2.0.3
openpyxl==3.0.9
packaging==21.0
phonenumberslite==8.12.36
Pillow==8.4.0
pipdeptree==2.2.0
pycparser==2.21
pyparsing==3.0.1
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
tablib==3.0.0
telepath==0.2
urllib3==1.26.7
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.4.5
xlwt==1.3.0
